requirements.txt
accelerate==1.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.0
bm25s==0.2.3
certifi==2024.8.30
charset-normalizer==3.4.0
datasets==3.0.1
diffusers==0.31.0
dill==0.3.8
faiss-gpu==1.7.2
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.6.1
huggingface-hub==0.26.1
idna==3.10
importlib-metadata==8.5.0
jinja2==3.1.4
loguru==0.7.2
markupsafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pipeline==0.1.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.29.1
psutil==6.1.0
pyarrow==17.0.0
pystemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.6.0
six==1.16.0
soupsieve==2.6
stemmer==0.0.4
sympy==1.13.1
sympy==1.13.3
-e file:///var/autofs/cl/home2/shintaro/text_simplification_for_image_generation
tokenizers==0.20.1
tomli==2.0.2
torch==2.2.2
torch==2.5.1
tqdm==4.66.5
transformers==4.45.2
triton==2.2.0
triton==3.1.0
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xformers==0.0.28.post3
xxhash==3.5.0
yapf==0.40.2
yarl==1.16.0
zipp==3.20.2
